+++ tests/LCacheTest.php
@@ -134,11 +134,7 @@ class LCacheTest extends \PHPUnit_Extensions_Database_TestCase
         $this->assertEquals($pool1->getLastAppliedEventID(), $pool2->getLastAppliedEventID());
     }
 
-    public function testAPCuL1Tombstone()
-    {
-        $l1 = new APCuL1('testAPCuL1Tombstone');
-        $this->performTombstoneTest($l1);
-    }
+
 
 
     protected function performTaggedSynchronizationTest($central, $first_l1, $second_l1)
@@ -326,19 +322,11 @@ class LCacheTest extends \PHPUnit_Extensions_Database_TestCase
         $l2 = new DatabaseL2($this->dbh);
     }
 
-    public function testExistsAPCuL1()
-    {
-        $l1 = new APCuL1('first');
-        $this->performExistsTest($l1);
-    }
-
     public function testExistsStaticL1()
     {
         $l1 = new StaticL1();
         $this->performExistsTest($l1);
     }
-
-
 
     public function testExistsIntegrated()
     {
@@ -360,31 +348,6 @@ class LCacheTest extends \PHPUnit_Extensions_Database_TestCase
         $myaddr = new Address('mybin', 'mykey');
         $l2->set('mypool', $myaddr, 'myvalue', null, ['mytag']);
         $this->assertEquals('myvalue', $l2->get($myaddr));
-    }
-
-    public function testAPCuL1PoolIDs()
-    {
-        // Test unique ID generation.
-        $l1 = new APCuL1();
-        $this->assertNotNull($l1->getPool());
-
-        // Test host-based generation.
-        $_SERVER['SERVER_ADDR'] = 'localhost';
-        $_SERVER['SERVER_PORT'] = '80';
-        $l1 = new APCuL1();
-        $this->assertEquals('localhost-80', $l1->getPool());
-    }
-
-    public function testAPCuL1Antirollback()
-    {
-        $l1 = new APCuL1('first');
-        $this->performL1AntirollbackTest($l1);
-    }
-
-    public function testAPCuL1HitMiss()
-    {
-        $l1 = new APCuL1('testAPCuL1HitMiss');
-        $this->performL1HitMissTest($l1);
     }
 
     public function testPoolIntegrated()
@@ -526,10 +489,7 @@ class LCacheTest extends \PHPUnit_Extensions_Database_TestCase
         $this->performHitSetCounterTest(new StaticL1());
     }
 
-    public function testAPCuL1Counters()
-    {
-        $this->performHitSetCounterTest(new APCuL1('counters'));
-    }
+
 
 
     public function testStaticL1ExcessiveOverheadSkipping()
@@ -537,31 +497,7 @@ class LCacheTest extends \PHPUnit_Extensions_Database_TestCase
         $this->performExcessiveOverheadSkippingTest(new StaticL1());
     }
 
-    public function testAPCuL1ExcessiveOverheadSkipping()
-    {
-        $this->performExcessiveOverheadSkippingTest(new APCuL1('overhead'));
-    }
 
-    public function testAPCuL1Expiration()
-    {
-        $l1 = new APCuL1();
-        $pool = new Integrated($l1, new StaticL2());
-        $myaddr = new Address('mybin', 'mykey');
-        $pool->set($myaddr, 'value', 1);
-        $this->assertEquals('value', $pool->get($myaddr));
-        $this->assertEquals($_SERVER['REQUEST_TIME'] + 1, $l1->getEntry($myaddr)->expiration);
-
-        // Setting items with past expirations should result in a nothing stored.
-        $myaddr2 = new Address('mybin', 'mykey2');
-        $l1->set(0, $myaddr2, 'value', $_SERVER['REQUEST_TIME'] - 1);
-        $this->assertNull($l1->get($myaddr2));
-
-        // Setting an TTL/expiration more than request time should be treated
-        // as an expiration.
-        $pool->set($myaddr, 'value', $_SERVER['REQUEST_TIME'] + 1);
-        $this->assertEquals('value', $pool->get($myaddr));
-        $this->assertEquals($_SERVER['REQUEST_TIME'] + 1, $l1->getEntry($myaddr)->expiration);
-    }
 
     public function testDatabaseL2BatchDeletion()
     {
@@ -583,4 +519,16 @@ class LCacheTest extends \PHPUnit_Extensions_Database_TestCase
     {
         return new \PHPUnit_Extensions_Database_DataSet_DefaultDataSet();
     }
+
+
+
+
+
+
+
+
+
+
+
+
 }
